Replace variable with direct expression
get_post is named clearly enough that it doesn't need to be assigned to
an explaining variable.

Omitting the variable means there's one less name in scope during the
rest of the function, which makes it simpler to read. There's no more
waiting for the other shoe to drop - "when will we see $post again?"